requirements.txt
# 基础库
torch>=1.10.0
numpy>=1.19.2
tqdm>=4.62.3

# 深度学习和训练
transformers>=4.37.0
tensorboard>=2.11.2

# 自然语言处理
tokenizers>=0.11.0
sentencepiece>=0.1.96

# 音频处理
vosk>=0.3.42
pydub>=0.25.1
edge-tts>=5.0.0
websockets>=10.3

# API和网络请求
requests>=2.27.1
python-dotenv>=0.19.2

# GUI相关
wxPython>=4.1.1
panda3d>=1.10.9

# 数据处理
json5>=0.9.6 

# 图像处理
Pillow>=8.0.0

# 日志
loguru>=0.6.0